API restructure in app/views/errors submodules.
