+++ controller/api/views/algorithms.py
@@ -1,26 +1,23 @@
 import logging
 import traceback
 
-from quart import Quart, request
+from quart import request, Blueprint
 
 from controller.api.DTOs.AlgorithmSpecificationsDTOs import AlgorithmDTO, AlgorithmSpecifications
-from controller.api.errors import BadRequest, BadUserInput
+from controller.api.errors.exceptions import BadRequest, BadUserInput
 from controller.api.services.run_algorithm import run_algorithm
 
-app = Quart(__name__)
+algorithms = Blueprint('algorithms_endpoint', __name__)
 
 
-# TODO break into views/app/errors
-
-
-@app.route("/algorithms")
+@algorithms.route("/algorithms")
 async def get_algorithms() -> str:
     algorithm_specifications = AlgorithmSpecifications().algorithms_list
 
     return AlgorithmDTO.schema().dumps(algorithm_specifications, many=True)
 
 
-@app.route("/algorithms/<algorithm_name>", methods=['POST'])
+@algorithms.route("/algorithms/<algorithm_name>", methods=['POST'])
 async def post_algorithm(algorithm_name: str) -> str:
     logging.info(f"Algorithm execution with name {algorithm_name}.")
 
@@ -36,13 +33,3 @@ async def post_algorithm(algorithm_name: str) -> str:
                          "Please inform the system administrator or try again later.")
 
     return response
-
-
-@app.errorhandler(BadRequest)
-def handle_bad_request(error: BadRequest):
-    return error.message, error.status_code
-
-
-@app.errorhandler(BadUserInput)
-def handle_bad_user_input(error: BadUserInput):
-    return error.message, error.status_code
